# sql: improve func expression type checking performance