--- a/pkg/sql/catalog/schemadesc/schema_desc.go
+++ b/pkg/sql/catalog/schemadesc/schema_desc.go
@@ -492,7 +492,7 @@ func (desc *immutable) GetResolvedFuncDefinition(
 	}
 	funcDef := &tree.ResolvedFunctionDefinition{
 		Name:      name,
-		Overloads: make([]tree.QualifiedOverload, 0, len(funcDescPb.Overloads)),
+		Overloads: make([]*tree.QualifiedOverload, 0, len(funcDescPb.Overloads)),
 	}
 	for i := range funcDescPb.Overloads {
 		retType := funcDescPb.Overloads[i].ReturnType