--- a/pkg/sql/catalog/seqexpr/sequence.go
+++ b/pkg/sql/catalog/seqexpr/sequence.go
@@ -70,16 +70,17 @@ func GetSequenceFromFunc(funcExpr *tree.FuncExpr) (*SeqIdentifier, error) {
 		found := false
 		for i := range def.Overloads {
 			// Find the overload that matches funcExpr.
-			if len(funcExpr.Exprs) == def.Overloads[i].Types.Length() {
+			ol := def.Overloads[i].(*tree.QualifiedOverload)
+			if len(funcExpr.Exprs) == ol.Types.Length() {
 				found = true
-				argTypes, ok := def.Overloads[i].Types.(tree.ArgTypes)
+				argTypes, ok := ol.Types.(tree.ArgTypes)
 				if !ok {
 					panic(pgerror.Newf(
 						pgcode.InvalidFunctionDefinition,
 						"%s has invalid argument types", funcExpr.Func.String(),
 					))
 				}
-				for i := 0; i < def.Overloads[i].Types.Length(); i++ {
+				for i := 0; i < ol.Types.Length(); i++ {
 					// Find the sequence name arg.
 					argName := argTypes[i].Name
 					if argName == builtinconstants.SequenceNameArg {