--- a/pkg/sql/sem/tree/type_check.go
+++ b/pkg/sql/sem/tree/type_check.go
@@ -1084,7 +1084,7 @@ func (expr *FuncExpr) TypeCheck(
 	var calledOnNullInputFns []overloadImpl
 	var notCalledOnNullInputFns []overloadImpl
 	for _, f := range fns {
-		if f.(QualifiedOverload).CalledOnNullInput {
+		if f.(*QualifiedOverload).CalledOnNullInput {
 			calledOnNullInputFns = append(calledOnNullInputFns, f)
 		} else {
 			notCalledOnNullInputFns = append(notCalledOnNullInputFns, f)
@@ -2936,7 +2936,7 @@ func (stripFuncsVisitor) VisitPost(expr Expr) Expr { return expr }
 // of QualifiedOverload. Also, the input should not be empty.
 func getMostSignificantOverload(
 	overloads []overloadImpl, searchPath SearchPath, expr *FuncExpr, getFuncSig func() string,
-) (QualifiedOverload, error) {
+) (*QualifiedOverload, error) {
 	ambiguousError := func() error {
 		return pgerror.Newf(
 			pgcode.AmbiguousFunction,
@@ -2947,13 +2947,13 @@ func getMostSignificantOverload(
 			formatCandidates(expr.Func.String(), overloads),
 		)
 	}
-	checkAmbiguity := func(oImpls []overloadImpl) (QualifiedOverload, error) {
+	checkAmbiguity := func(oImpls []overloadImpl) (*QualifiedOverload, error) {
 		if len(oImpls) > 1 {
 			// Throw ambiguity error if there are more than one candidate overloads from
 			// same schema.
-			return QualifiedOverload{}, ambiguousError()
+			return nil, ambiguousError()
 		}
-		return oImpls[0].(QualifiedOverload), nil
+		return oImpls[0].(*QualifiedOverload), nil
 	}
 
 	if searchPath == nil || searchPath == EmptySearchPath {
@@ -2963,10 +2963,10 @@ func getMostSignificantOverload(
 	udfFound := false
 	uniqueSchema := true
 	for i, o := range overloads {
-		if o.(QualifiedOverload).IsUDF {
+		if o.(*QualifiedOverload).IsUDF {
 			udfFound = true
 		}
-		if i > 0 && o.(QualifiedOverload).Schema != overloads[i-1].(QualifiedOverload).Schema {
+		if i > 0 && o.(*QualifiedOverload).Schema != overloads[i-1].(*QualifiedOverload).Schema {
 			uniqueSchema = false
 		}
 	}
@@ -2987,16 +2987,16 @@ func getMostSignificantOverload(
 	}
 
 	found := false
-	var ret QualifiedOverload
+	var ret *QualifiedOverload
 	for i, n := 0, searchPath.NumElements(); i < n; i++ {
 		schema := searchPath.GetSchema(i)
 		for i := range overloads {
-			if overloads[i].(QualifiedOverload).Schema == schema {
+			if overloads[i].(*QualifiedOverload).Schema == schema {
 				if found {
-					return QualifiedOverload{}, ambiguousError()
+					return nil, ambiguousError()
 				}
 				found = true
-				ret = overloads[i].(QualifiedOverload)
+				ret = overloads[i].(*QualifiedOverload)
 			}
 		}
 		if found {
@@ -3007,7 +3007,7 @@ func getMostSignificantOverload(
 		// This should never happen. Otherwise, it means we get function from a
 		// schema no on the given search path or we try to resolve a function on an
 		// explicit schema, but get some function from other schemas are fetched.
-		return QualifiedOverload{}, pgerror.Newf(pgcode.UndefinedFunction, "unknown signature: %s", getFuncSig())
+		return nil, pgerror.Newf(pgcode.UndefinedFunction, "unknown signature: %s", getFuncSig())
 	}
 	return ret, nil
 }