2 commits:
(1) change overloads returned from function resolver to pointers to avoid value copying.
(2) expand the `overloadImpl` interface to allow accessing function overload properties without type casting. This allows us to get rid of the slice allocation in type checking.

All these would be able to improve the type checking performance by 20% cpu time overall:
```
name                                                       old time/op    new time/op    delta
FuncExprTypeCheck/builtin_called_on_null_input-24            1.05µs ± 1%    0.68µs ± 1%  -35.53%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_not_called_on_null_input-24        1.00µs ± 1%    0.66µs ± 1%  -34.04%  (p=0.000 n=9+10)
FuncExprTypeCheck/builtin_aggregate-24                       2.17µs ± 0%    1.06µs ± 0%  -51.07%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24    1.23µs ± 1%    0.80µs ± 1%  -35.02%  (p=0.000 n=10+9)
FuncExprTypeCheck/udf_same_name_as_builtin-24                1.50µs ± 0%    0.91µs ± 1%  -39.53%  (p=0.000 n=10+9)
FuncExprTypeCheck/udf_across_different_schemas-24            2.43µs ± 0%    1.80µs ± 1%  -25.90%  (p=0.000 n=8+10)
FuncExprTypeCheck/unary_operator-24                           344ns ± 2%     349ns ± 2%     ~     (p=0.065 n=10+10)
FuncExprTypeCheck/binary_operator-24                         1.36µs ± 0%    1.35µs ± 0%   -0.72%  (p=0.000 n=10+10)
FuncExprTypeCheck/comparison_operator-24                     2.16µs ± 0%    2.18µs ± 0%   +0.81%  (p=0.000 n=10+10)
FuncExprTypeCheck/tuple_comparison_operator-24               6.45µs ± 0%    6.47µs ± 1%     ~     (p=0.239 n=10+10)
FuncExprTypeCheck/tuple_in_operator-24                       1.13µs ± 1%    1.14µs ± 0%   +1.21%  (p=0.000 n=10+9)

name                                                       old alloc/op   new alloc/op   delta
FuncExprTypeCheck/builtin_called_on_null_input-24              132B ± 6%       34B ±10%  -74.53%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_not_called_on_null_input-24          129B ± 0%       33B ± 2%  -74.81%  (p=0.000 n=8+8)
FuncExprTypeCheck/builtin_aggregate-24                         467B ± 0%       73B ± 0%  -84.37%  (p=0.000 n=9+8)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24      145B ± 0%       32B ± 0%  -77.93%  (p=0.000 n=8+8)
FuncExprTypeCheck/udf_same_name_as_builtin-24                  273B ± 0%       64B ± 1%  -76.53%  (p=0.000 n=9+8)
FuncExprTypeCheck/udf_across_different_schemas-24              674B ± 0%      465B ± 0%  -30.90%  (p=0.000 n=8+9)
FuncExprTypeCheck/unary_operator-24                           0.00B          0.00B          ~     (all equal)
FuncExprTypeCheck/binary_operator-24                           113B ± 0%      113B ± 1%     ~     (p=0.214 n=6+9)
FuncExprTypeCheck/comparison_operator-24                       153B ± 0%      155B ± 4%     ~     (p=0.793 n=9+9)
FuncExprTypeCheck/tuple_comparison_operator-24                 894B ± 0%      892B ± 0%     ~     (p=0.057 n=9+8)
FuncExprTypeCheck/tuple_in_operator-24                         400B ± 0%      400B ± 0%     ~     (p=1.161 n=8+9)

name                                                       old allocs/op  new allocs/op  delta
FuncExprTypeCheck/builtin_called_on_null_input-24              7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_not_called_on_null_input-24          7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate-24                         16.0 ± 0%       4.0 ± 0%  -75.00%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24      8.00 ± 0%      3.00 ± 0%  -62.50%  (p=0.000 n=10+10)
FuncExprTypeCheck/udf_same_name_as_builtin-24                  11.0 ± 0%       4.0 ± 0%  -63.64%  (p=0.000 n=10+10)
FuncExprTypeCheck/udf_across_different_schemas-24              16.0 ± 0%       9.0 ± 0%  -43.75%  (p=0.002 n=8+10)
FuncExprTypeCheck/unary_operator-24                            0.00           0.00          ~     (all equal)
FuncExprTypeCheck/binary_operator-24                           5.00 ± 0%      5.00 ± 0%     ~     (all equal)
FuncExprTypeCheck/comparison_operator-24                       5.00 ± 0%      5.00 ± 0%     ~     (all equal)
FuncExprTypeCheck/tuple_comparison_operator-24                 19.0 ± 0%      19.0 ± 0%     ~     (all equal)
FuncExprTypeCheck/tuple_in_operator-24                         7.00 ± 0%      7.00 ± 0%     ~     (all equal)
```