@chengxiong-ruan 2022-10-10
@ajwerner not sure you hate it less or not. But I went to the direction of expanding the `overloadImpl` interface so that we can access several function properties without doing type assertion. To avoid value copy, I also make `Overload` methods to have pointer receiver. By doing all of these, it seemed that the performance was improve even greater than previous change by looking the type checking benchmark. (20% improvement -> 35% improvement in most cases) also good to see operators are not affected.