@ajwerner 2022-10-12
I didn't love it. I have a refactoring which I think I like more. I spent a little longer on it than I'd like to admit. See how it makes you feel. 	https://github.com/cockroachdb/cockroach/pull/89832