@chengxiong-ruan 2022-10-12
> I didn't love it. I have a refactoring which I think I like more. I spent a little longer on it than I'd like to admit. See how it makes you feel. #89832

It looks good. Though I feel like the complexity grows a bit, it's definitely much cleaner than my kinda ugly solution here. I'm also ok with slightly less cpu gain since I'm super happy with the much better allocation gain. Thanks for working on this! Closing this PR in favor of it.