@ajwerner 2022-10-12
The CPU gains are not clear. Did you rebase yours? Mine beat the old version of yours strictly when I pulled it yesterday. Probably it's not worth agonizing over the nanoseconds anymore on this one.